Add auth check for platform events_client.py
Add reject_event function

Add some logging statements

Add default logging handler removal

+++ server/platform_event_queue_processor.py
@@ -1,4 +1,7 @@
 import asyncio
+import logging
+
+import security
 
 from models import tenant
 from tenant import processor
@@ -7,9 +10,18 @@ async def process_events(async_queue: asyncio.Queue):
     while True:
         tenant_event: tenant.TenantEvent = await async_queue.get()
 
-        if tenant_event.type == 'create':
-            processor.create_tenant(tenant_event=tenant_event)
-        elif tenant_event.type == 'status':
-            processor.status_check(tenant_event)
-        elif tenant_event.type == 'delete':
-            processor.delete_tenant(tenant_event)
+        logging.debug('Event item retreived from queue.')
+        if security.platform_auth.check_platform_event_authorized(tenant_event):
+            if tenant_event.type == 'create':
+                processor.create_tenant(tenant_event=tenant_event)
+            elif tenant_event.type == 'status':
+                processor.status_check(tenant_event)
+            elif tenant_event.type == 'delete':
+                processor.delete_tenant(tenant_event)
+            else:
+                processor.reject_event(tenant_event, reason="invalid_event_type")
+        else:
+            processor.reject_event(tenant_event, reason="invalid_event_auth")
+
+        async_queue.task_done()
+        logging.debug('Event item fully processed.')
